+++ bloc_flutter/lib/main_local_storage.dart
@@ -8,7 +8,7 @@ import 'package:bloc_flutter_sample/main.dart' as app;
 import 'package:blocs/blocs.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:todos_repository_core/todos_repository_core.dart';
-import 'package:todos_repository_flutter/todos_repository_flutter.dart';
+import 'package:todos_repository_simple/todos_repository_simple.dart';
 
 void main() {
   app.main(
