+++ inherited_widget/pubspec.yaml
@@ -5,8 +5,8 @@ environment:
   sdk: '>=2.0.0 <3.0.0'
 
 dependencies:
-  todos_repository_flutter:
-    path: ../todos_repository_flutter
+  todos_repository_simple:
+    path: ../todos_repository_simple
   todos_app_core:
     path: ../todos_app_core
   flutter:
