+++ mvc/lib/src/todo_list_model.dart
@@ -6,7 +6,7 @@ import 'dart:async';
 
 import 'package:path_provider/path_provider.dart';
 import 'package:todos_repository_core/todos_repository_core.dart';
-import 'package:todos_repository_flutter/todos_repository_flutter.dart';
+import 'package:todos_repository_simple/todos_repository_simple.dart';
 import 'package:mvc/src/models.dart';
 
 class TodoListModel {
