+++ mvc/pubspec.yaml
@@ -15,8 +15,8 @@ dependencies:
     path: ../scoped_model
   todos_app_core:
     path: ../todos_app_core
-  todos_repository_flutter:
-    path: ../todos_repository_flutter
+  todos_repository_simple:
+    path: ../todos_repository_simple
   mvc_pattern:
 
 dev_dependencies:
