+++ redux/pubspec.yaml
@@ -12,8 +12,8 @@ dependencies:
     sdk: flutter
   todos_app_core:
     path: ../todos_app_core
-  todos_repository_flutter:
-    path: ../todos_repository_flutter
+  todos_repository_simple:
+    path: ../todos_repository_simple
 
 dev_dependencies:
   test: ^1.3.0
