+++ scoped_model/lib/main.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:scoped_model_sample/app.dart';
-import 'package:todos_repository_flutter/todos_repository_flutter.dart';
+import 'package:todos_repository_simple/todos_repository_simple.dart';
 
 void main() {
   var todoRepo = const TodosRepositoryFlutter(
