+++ todos_repository_simple/lib/todos_repository_simple.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the MIT license that can be found
 // in the LICENSE file.
 
-library todos_repository_flutter;
+library todos_repository_simple;
 
 export 'src/file_storage.dart';
 export 'src/reactive_repository.dart';
