+++ todos_repository_simple/pubspec.yaml
@@ -1,4 +1,4 @@
-name: todos_repository_flutter
+name: todos_repository_simple
 description: An app-agnostic data source that can be used by all implementations
 
 environment:
