+++ todos_repository_simple/test/file_storage_test.dart
@@ -6,7 +6,7 @@ import 'dart:io';
 
 import 'package:test/test.dart';
 import 'package:todos_repository_core/todos_repository_core.dart';
-import 'package:todos_repository_flutter/todos_repository_flutter.dart';
+import 'package:todos_repository_simple/todos_repository_simple.dart';
 
 main() {
   group('FileStorage', () {
