+++ vanilla/lib/main.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:todos_repository_flutter/todos_repository_flutter.dart';
+import 'package:todos_repository_simple/todos_repository_simple.dart';
 import 'package:vanilla/app.dart';
 
 void main() {
