+++ vanilla/pubspec.yaml
@@ -9,8 +9,8 @@ dependencies:
     sdk: flutter
   todos_app_core:
     path: ../todos_app_core
-  todos_repository_flutter:
-    path: ../todos_repository_flutter
+  todos_repository_simple:
+    path: ../todos_repository_simple
 
 dev_dependencies:
   mockito: ^3.0.0
